# Cleanup for more readable + benchmark ExtractStacktrace