In performance nothing change
```bash
benchstat old.txt new.txt
```
```text
name                  old time/op    new time/op    delta
ExtractStacktrace-12    6.73µs ±17%    6.59µs ±16%   ~     (p=0.057 n=98+98)

name                  old alloc/op   new alloc/op   delta
ExtractStacktrace-12    2.56kB ± 0%    2.56kB ± 0%   ~     (all equal)

name                  old allocs/op  new allocs/op  delta
ExtractStacktrace-12      26.0 ± 0%      26.0 ± 0%   ~     (all equal)
```